backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "bcrypt<5.0.0,>=4.0.0",
    "email-validator>=2.0.0",
    "fastapi>=0.124.4",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
